+++ .docusaurus/@easyops-cn/docusaurus-search-local/default/generated-constants.js
@@ -1,6 +1,6 @@
 import lunr from "/Users/bdan/Projects/SAFROCHAIN/node/Docs/node_modules/lunr/lunr.js";
 export const removeDefaultStopWordFilter = ["en"];
 export const language = ["en"];
-export const searchIndexUrl = "search-index{dir}.json?_=9141cf41";
+export const searchIndexUrl = "search-index{dir}.json?_=b67f9186";
 export const searchResultLimits = 10;
 export const fuzzyMatchingDistance = 1;